Hide the platform model logic from the python backend implementation
