+++ CMakeLists.txt
@@ -206,8 +206,6 @@ set(
   src/pb_stub.cc
   src/pb_response_iterator.h
   src/pb_response_iterator.cc
-  src/model_context.h
-  src/model_context.cc
 )
 
 list(APPEND
