+++ src/pb_stub.h
@@ -49,7 +49,6 @@
 #include "message_queue.h"
 #include "metric.h"
 #include "metric_family.h"
-#include "model_context.h"
 #include "pb_log.h"
 #include "pb_response_iterator.h"
 #include "pb_utils.h"
@@ -161,6 +160,30 @@ class LogMessage {
 };
 
 #define LOG_FL(FN, LN, LVL) LogMessage((char*)(FN), LN, LVL).stream()
+
+
+class ModelScanner {
+ public:
+  // Scans and establishes path for serving the python model.
+  void Init(
+      const std::string& model_path, const std::string& platform,
+      const std::string& triton_install_path, const std::string& model_version);
+  // Sets up the python stub with appropriate paths.
+  void StubSetup(py::module& sys);
+
+  std::string& PythonModelPath() { return python_model_path_; }
+  std::string& ModelPath() { return model_path_; }
+
+ private:
+  std::string python_model_path_;
+  std::string model_path_;
+  std::string model_version_;
+  std::string python_backend_folder_;
+  std::string platform_model_;
+
+  enum ModelType { DEFAULT, PLATFORM };
+  ModelType type_;
+};
 
 // The payload for the stub_to_parent message queue. This struct serves as a
 // wrapper for different types of messages so that they can be sent through the
@@ -315,7 +338,7 @@ class Stub {
   bi::interprocess_mutex* parent_mutex_;
   bi::interprocess_condition* parent_cond_;
   bi::interprocess_mutex* health_mutex_;
-  ModelContext model_context_;
+  ModelScanner model_scanner_;
   std::string model_version_;
   std::string name_;
   std::string model_path_;
